+++ CHANGELOG.md
@@ -21,3 +21,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a change log
 
 ### Removed
+
+- Removed global app loading state